repository classes updated, data generator fully reworked

+++ ConsoleApplication/Comment.cs
@@ -1,16 +1,13 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+
 
 namespace ConsoleApplication
 {
     class Comment
     {
-        public int id;
-        public int authorId;
-        public int postId;
+        public long id;
+        public long authorId;
+        public long postId;
         public string text;
         public DateTime publishTime;
         public bool isPinned;
